Added support for status bar color, fault reporting, and login page.
Status Bar color has been changed to a variation of the color scheme to
make the overall look more cohesive.

Fault Reporting Forms allow users to submit faults to the servers, and
view a submission page.

Login Page has password masking that can be turned off, disabled navigation
drawer for users who are not logged in.